requirements.txt
einops
kaggle
matplotlib
monai==1.5.2
numpy
Pillow
rasterio
scikit-learn
torch
tqdm